[fix] Robust conversation handling for chat routes (not fully working yet)
Refactor app/api/chat/save/route.ts to better distinguish new vs. existing chat sessions, aiming for correct conversation creation and message updates.
Improve logging for conversation fetch/save, including message counts and error fallback handling.
Remove unused gmailMutations import from convex/_generated/api.d.ts.
Adjust argument order in convex/chat.ts for consistency in conversation and message mutations.
⚠️ Known issue: All messages are still being saved to a single conversation; proper separation is not fully working yet.
