+++ app/(dashboard)/_components/chat/hooks/useChat.ts
@@ -1,0 +1,143 @@
+import { useState, useCallback } from 'react'
+import { Message } from '@/app/types/chat'
+import { sendChatMessage } from '../utils/api-utils'
+import { ChatStateReturnType } from './useChatState'
+
+export const useChat = (
+  chatState: ChatStateReturnType
+) => {
+  const {
+    sessionId,
+    setSessionId,
+    messages,
+    setMessages,
+    setConversationSaved,
+    isLoading,
+    setIsLoading,
+    error,
+    setError
+  } = chatState
+  const [referencedMessage, setReferencedMessage] = useState<Message | null>(null)
+
+  const handleSendMessage = useCallback(async (content: string) => {
+    if (!content || typeof content !== 'string' || !content.trim()) return;
+
+    const newMessage: Message = {
+      id: Date.now(),
+      content,
+      role: 'user',
+      timestamp: new Date().toISOString(),
+      referencedMessage: referencedMessage ? {
+        id: referencedMessage.id,
+        content: referencedMessage.content
+      } : undefined,
+      chat_response: content
+    }
+
+    try {
+      setIsLoading(true)
+      setError(null)
+
+      // Add user message and typing indicator
+      setMessages(prev => [
+        ...prev,
+        newMessage,
+        {
+          id: Date.now(),
+          content: '...',
+          role: 'assistant',
+          timestamp: new Date().toISOString(),
+          status: 'typing',
+          chat_response: ''
+        }
+      ]);
+
+      // Mark conversation as unsaved ONLY after user message is added
+      setConversationSaved(false);
+
+      // Always send the first message with sessionId=null and isFirstMessage=true
+      // When starting a new chat, ensure sessionId is null so backend creates a new conversation
+      const isFirstMessage = sessionId === null;
+      const backendSessionId = isFirstMessage ? null : sessionId;
+
+      // Send message to the backend
+      const data = await sendChatMessage(content, isFirstMessage, backendSessionId);
+
+      // Update messages with the response
+      setMessages(prev => {
+        const withoutTyping = prev.filter(msg => msg.status !== 'typing');
+        return [...withoutTyping, {
+          id: Date.now(),
+          content: data.chat_response,
+          chat_response: data.chat_response,
+          role: 'assistant',
+          timestamp: new Date().toISOString(),
+          suggestions: data.suggestions || []
+        }];
+      });
+
+      // Only update sessionId from backend (never generate a local one for persistence)
+      // Only set sessionId if we don't already have a valid one
+      const isValidBackendSession = typeof sessionId === 'string' && sessionId.match(/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/);
+      if (data.session_id && !isValidBackendSession) {
+        console.log('Received session ID from API:', data.session_id);
+        setSessionId(data.session_id);
+        // Do NOT set conversationSaved(false) here; already set above after user message
+      } else if (!sessionId) {
+        // If the backend fails to return a sessionId, do not attempt to save or persist
+        console.warn('No session ID received from API. Will not persist conversation until backend provides one.');
+      } // else: do not set conversationSaved(false) again
+
+      setReferencedMessage(null)
+
+    } catch (error) {
+      console.error('Failed to send message:', error)
+      setMessages(prev => prev.filter(msg => msg.status !== 'typing'))
+      setError((error as Error).message)
+    } finally {
+      setIsLoading(false)
+    }
+  }, [referencedMessage, sessionId, messages.length, setMessages, setSessionId, setConversationSaved, setIsLoading, setError])
+
+  const handleMessageReference = useCallback((message: Message) => {
+    setReferencedMessage(message)
+  }, [])
+
+  const handleClearReference = useCallback(() => {
+    setReferencedMessage(null)
+  }, [])
+
+  const handleOptionClick = useCallback((option: { text: string }) => {
+    if (!option?.text) return;
+    handleSendMessage(option.text);
+  }, [handleSendMessage]);
+
+  const handleFollowUpClick = useCallback((choice: string) => {
+    handleSendMessage(choice);
+  }, [handleSendMessage]);
+
+  const handleReferenceClick = useCallback((messageId: number) => {
+    const messageElement = document.getElementById(`message-${messageId}`);
+    if (messageElement) {
+      messageElement.scrollIntoView({ 
+        behavior: 'smooth',
+        block: 'center'
+      });
+      // Add a brief highlight effect
+      messageElement.classList.add('bg-yellow-100/50');
+      setTimeout(() => {
+        messageElement.classList.remove('bg-yellow-100/50');
+      }, 2000);
+    }
+  }, []);
+
+  return {
+    referencedMessage,
+    handleSendMessage,
+    handleMessageReference,
+    handleClearReference,
+    handleOptionClick,
+    handleFollowUpClick,
+    handleReferenceClick
+  }
+}
