+++ app/api/chat/save/route.ts
@@ -46,38 +46,128 @@ export async function POST(request: Request) {
     // Generate a title if not provided
     const chatTitle = title || `Chat from ${new Date().toLocaleString()}`;
 
-    // If we have an existing conversationId, add the latest message instead of creating a new conversation
-    if (conversationId) {
-      console.log(`[${requestId}] Updating existing conversation:`, {
-        conversationId,
+    // Use conversationId if provided (prioritize it over sessionId)
+    // If not, use sessionId as the conversationId to ensure continuity
+    const targetConversationId = conversationId || sessionId;
+
+    // Check if sessionId is from a new chat session
+    // If it begins with the backend's format (UUID), treat it as a potential new conversation
+    const isNewBackendSession = typeof sessionId === 'string' && 
+      sessionId.match(/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/);
+    
+    // Force a new conversation if it's a new backend session without an existing conversationId
+    if (isNewBackendSession && !conversationId) {
+      console.log(`[${requestId}] Detected new backend session ID, creating new conversation:`, {
+        sessionId,
         userId,
         messageCount: formattedMessages.length
       });
 
-      // Get the latest message only
-      const latestMessage = formattedMessages[formattedMessages.length - 1];
-      
-      if (latestMessage) {
-        const result = await convex.mutation(api.chat.addMessageToConversation, {
-          conversationId,
+      // Create a new conversation in Convex with the backend session ID
+      const newConversationId = await convex.mutation(api.chat.createConversation, {
+        userId,
+        title: chatTitle,
+        messages: formattedMessages
+      });
+
+      console.log(`[${requestId}] Successfully created new conversation for new session`, {
+        sessionId,
+        conversationId: newConversationId,
+        messageCount: formattedMessages.length
+      });
+
+      return NextResponse.json({
+        success: true,
+        conversationId: newConversationId,
+        metadata: {
+          request_id: requestId,
+          timestamp: new Date().toISOString(),
+          action: 'created'
+        }
+      });
+    }
+
+    // Check if we're updating an existing conversation
+    if (targetConversationId) {
+      console.log(`[${requestId}] Updating existing conversation:`, {
+        conversationId: targetConversationId,
+        userId,
+        messageCount: formattedMessages.length
+      });
+
+      try {
+        // First check if the conversation exists
+        const existingConversation = await convex.query(api.chat.getConversation, {
           userId,
-          message: latestMessage
+          conversationId: targetConversationId
         });
 
-        console.log(`[${requestId}] Successfully updated conversation`, {
-          conversationId,
-          result
-        });
+        if (!existingConversation) {
+          console.log(`[${requestId}] Conversation ID ${targetConversationId} not found, creating new conversation instead`);
+          // Continue to create new conversation below
+        } else {
+          // Get the latest message only
+          const latestUserMessage = [...formattedMessages].reverse().find(msg => msg.role === 'user');
+          const latestAssistantMessage = [...formattedMessages].reverse().find(msg => msg.role === 'assistant');
+          
+          // Update with the most recent messages if available
+          if (latestUserMessage && latestAssistantMessage) {
+            // Add user message first
+            await convex.mutation(api.chat.addMessageToConversation, {
+              conversationId: targetConversationId,
+              userId,
+              message: latestUserMessage
+            });
+            
+            // Then add assistant response
+            const result = await convex.mutation(api.chat.addMessageToConversation, {
+              conversationId: targetConversationId,
+              userId,
+              message: latestAssistantMessage
+            });
 
-        return NextResponse.json({
-          success: true,
-          conversationId,
-          metadata: {
-            request_id: requestId,
-            timestamp: new Date().toISOString(),
-            action: 'updated'
+            console.log(`[${requestId}] Successfully updated conversation with user & assistant messages`, {
+              conversationId: targetConversationId,
+              result
+            });
+
+            return NextResponse.json({
+              success: true,
+              conversationId: targetConversationId,
+              metadata: {
+                request_id: requestId,
+                timestamp: new Date().toISOString(),
+                action: 'updated'
+              }
+            });
+          } else if (latestUserMessage) {
+            // Add just the user message if that's all we have
+            const result = await convex.mutation(api.chat.addMessageToConversation, {
+              conversationId: targetConversationId,
+              userId,
+              message: latestUserMessage
+            });
+
+            console.log(`[${requestId}] Successfully updated conversation with user message only`, {
+              conversationId: targetConversationId,
+              result
+            });
+
+            return NextResponse.json({
+              success: true,
+              conversationId: targetConversationId,
+              metadata: {
+                request_id: requestId,
+                timestamp: new Date().toISOString(),
+                action: 'updated'
+              }
+            });
           }
-        });
+        }
+      } catch (error) {
+        console.error(`[${requestId}] Error checking/updating conversation:`, error);
+        console.log(`[${requestId}] Falling back to creating new conversation`);
+        // Continue to create new conversation
       }
     }
